Added list all endpoints and get one endpoint.
list all endpoints is using page-by-key pagination to grab all of the
endpoints.  get one simply grabs a single endpoint.  I tried to use
page-by-offset with endpoints as enough of the data is provided, but
it doesn't work.  If it did, it would be much faster
This commit closes out jira issues SSD-680 & SSD-681.