Added base app, helpers, etc
Added a ton here to supply the needful for the clients to come.   I have
been studying the go-github/github repo and used a lot of their
organization and even copied a few files directly.  Licenses should be
in place where i did that.  If not, they will be corrected when found.